Include signal.h in libc bindings
