+++ rtlib/src/main/resources/rt.cpp
@@ -3,6 +3,7 @@
 #include <gc.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <signal.h>
 #include <math.h>
 
 namespace __cxxabiv1 {
